pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codetree-rag"
version = "0.1.0"
description = "Vectorless RAG for Code Repositories - Navigate your codebase with LLM reasoning"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Tony", email = "1094086026@qq.com"}
]
keywords = ["rag", "code", "llm", "ai", "retrieval", "vectorless"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ollama = [
    "ollama>=0.1.0",
]

[project.scripts]
codetree = "codetree.cli:main"

[project.urls]
Homepage = "https://github.com/toller892/codetree"
Repository = "https://github.com/toller892/codetree"
Issues = "https://github.com/toller892/codetree/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
